updated test_general_agent to pytest
