+++ bankcraft/tests/test_general_agent.py
@@ -1,62 +1,82 @@
-from unittest import TestCase
-from bankcraft.model import Model
+import pandas as pd
+import pytest
+
 from bankcraft.agent.general_agent import GeneralAgent
+from bankcraft.model import Model
 from bankcraft.agent.bank import Bank
 from mesa.time import RandomActivation
 from mesa.datacollection import DataCollector
-import pandas as pd
+
+account_initial_balance = 1500
+num_banks = 1
+txn_amount = 300
+Model.schedule = RandomActivation(Model)
+Model.datacollector = \
+    DataCollector(tables={"transactions": ["sender", "receiver", "amount", "step", "txn_id",
+                                           "txn_type", "sender_account_type", "description"]})
 
 
-class TestGeneralAgent(TestCase):
-    def setUp(self):
-        self.model = Model
-        self.agent1 = GeneralAgent(self.model)
-        self.num_banks = 1
-        self.num_account_types = 3
-        self.account_initial_balance = 500
-        self.model.banks = [Bank(self.model) for _ in range(self.num_banks)]
-        self.agent1.bank_accounts = self.agent1.assign_bank_account(self.model, self.account_initial_balance)
-        self.agent2 = GeneralAgent(self.model)
-        self.model.schedule = RandomActivation(self.model)
-        self.model.datacollector = \
-            DataCollector(tables={"transactions": ["sender", "receiver", "amount", "step", "txn_id",
-                                                   "txn_type", "sender_account_type", "description"]})
+@pytest.fixture
+def banks():
+    Model.banks = [Bank(Model) for _ in range(num_banks)]
+    return Model.banks
 
-    def test_assign_bank_account(self):
-        self.assertIsNotNone(self.agent1.bank_accounts)
 
-    def test_update_wealth(self):
-        self.assertEqual(self.agent2.wealth, 0)
-        self.agent2.bank_accounts = self.agent2.assign_bank_account(self.model, self.account_initial_balance)
-        self.agent2.update_wealth()
-        self.assertEqual(self.agent2.wealth, self.num_banks*self.num_account_types*self.account_initial_balance)
+@pytest.fixture
+def agent():
+    return GeneralAgent(Model)
 
-    def test_pay(self):
-        self.agent2.bank_accounts = self.agent2.assign_bank_account(self.model, self.account_initial_balance)
-        self.agent1.pay(300, self.agent2, "cash", "gift")
-        self.assertEqual(self.agent1.wealth, 1200)
-        self.assertEqual(self.agent2.wealth, 1800)
 
-        with self.assertRaises(ValueError):
-            self.agent1.pay(300, self.agent2, "cash", "gift")
+@pytest.fixture
+def other_agent():
+    return GeneralAgent(Model)
 
-        with self.assertRaises(Exception):
-            self.agent1.pay(100, self.agent2, "e-transfer", "gift")
 
-    def test_update_records(self):
-        self.agent1.update_records(self.agent2, 100, "cheque", "debt")
-        self.model.datacollector.collect(self.model)
-        expected_transaction_data = {
-            "sender": self.agent1.unique_id,
-            "receiver": self.agent2.unique_id,
-            "amount": 100,
-            "step": 0,
-            "txn_id": f"{str(self.agent1.unique_id)}_{str(self.agent1.txn_counter)}",
-            "txn_type": "cheque",
-            "description": "debt",
-        }
-        actual_transaction_data = self.model.datacollector.get_table_dataframe("transactions").iloc[0]
-        pd.testing.assert_series_equal(pd.Series(expected_transaction_data, name=0), actual_transaction_data)
+def test_bank_account_is_not_none_after_assigning(agent, banks):
+    agent.bank_accounts = agent.assign_bank_account(Model, account_initial_balance)
+    assert agent.bank_accounts is not None
 
-    if __name__ == '__main__':
-        TestCase.main()
+
+def test_update_wealth_after_assigning_bank_account(agent, banks):
+    agent_default_wealth = agent.wealth
+    agent.bank_accounts = agent.assign_bank_account(Model, account_initial_balance)
+    agent.update_wealth()
+    assert (agent_default_wealth == 0 and
+            agent.wealth == num_banks*account_initial_balance)
+
+
+def test_pay_changes_chequing_balance(banks, agent, other_agent):
+    agent.bank_accounts = agent.assign_bank_account(Model, account_initial_balance)
+    other_agent.bank_accounts = other_agent.assign_bank_account(Model, account_initial_balance)
+    agent.pay(txn_amount, other_agent, "cash", "gift")
+    assert (agent.bank_accounts[0][0].balance == account_initial_balance - txn_amount and
+            other_agent.bank_accounts[0][0].balance == account_initial_balance + txn_amount)
+
+
+def test_undefined_tnx_type_does_not_change_wealth(banks, agent, other_agent):
+    agent.bank_accounts = agent.assign_bank_account(Model, account_initial_balance)
+    agent.update_wealth()
+    pre_txn_wealth = agent.wealth
+    other_agent.bank_accounts = other_agent.assign_bank_account(Model, account_initial_balance)
+    agent.pay(txn_amount, other_agent, "e-transfer", "gift")
+    assert pre_txn_wealth == agent.wealth
+
+
+def test_updating_txn_records(agent, other_agent):
+    Model.datacollector = \
+        DataCollector(tables={"transactions": ["sender", "receiver", "amount", "step", "txn_id",
+                                               "txn_type", "sender_account_type", "description"]})
+    agent.update_records(other_agent, txn_amount, "cheque", "chequing", "debt")
+    Model.datacollector.collect(Model)
+    expected_txn_data = {
+        "sender": agent.unique_id,
+        "receiver": other_agent.unique_id,
+        "amount": txn_amount,
+        "step": 0,
+        "txn_id": f"{str(agent.unique_id)}_{str(agent.txn_counter)}",
+        "txn_type": "cheque",
+        "sender_account_type": "chequing",
+        "description": "debt",
+    }
+    model_txn_data = Model.datacollector.get_table_dataframe("transactions").iloc[0]
+    pd.testing.assert_series_equal(pd.Series(expected_txn_data, name=0), model_txn_data)
